requirements.txt
openai
langchain
langchain-community
langchain-openai
langchain-text-splitters
chromadb
streamlit
python-dotenv
tiktoken
pypdf